rag-service/requirements.txt
Flask==2.3.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
openai==0.28.0
python-dotenv==1.0.0
huggingface_hub==0.20.3  # Use a compatible version
torch==2.0.1  # Add PyTorch for sentence-transformers
transformers==4.28.1  # Add transformers for sentence-transformers